requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
PyMuPDF==1.22.5
nltk==3.6.3
openai==0.27.8
python-dotenv==0.19.0
Flask-Caching==2.0.2 